pybibiocr/onnx/cnocr_lite/requirements.txt
tornado==5.1.1
numpy==1.19.1
opencv_python==4.3.0.36
onnxruntime==1.4.0
Shapely==1.7.0
pyclipper==1.2.0
Pillow==7.2.0
LAC==2.1.2
